Documentation included in the commit
It's usually bad practice to include automatically-generated files in
the commit, but since the docs here explain how to install the Chicken
compiler and how to generate the executable/docs, I thought it is
better to have it available in the repository. In order to keep it
updated, I modified the Makefile so that it automatically regenerates
it any time `make` is called.